api/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
scikit-learn>=1.4,<2.0
spacy>=3.7,<4.0
pdfplumber>=0.11,<1.0
python-docx>=1.1,<2.0
python-multipart>=0.0.9,<1.0
